Create tab screens
Also moved everything from App.tsx into HomeScreen.tsx component
