+++ package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.17.7",
+    "@react-navigation/bottom-tabs": "^6.3.2",
+    "@react-navigation/native": "^6.0.11",
     "expo": "~45.0.0",
     "expo-status-bar": "~1.3.0",
     "react": "17.0.2",
